+++ lib/constants/api.dart
@@ -2,8 +2,12 @@ class API {
   static const Object DEFAULT_HEADERS = {
     'Accept': 'application/json',
   };
-  static const String GET_RANDOM_USER_LIST_URL = "https://randomuser.me/api"; //https://randomuser.me/api
-  static const String GET_RESERVATION_LIST_URL = "https://backend.rentify.rbit.pe/api/reservations"; //http://localhost:5000/api/reservations - https://localhost:5001/api/reservations
-  static const String GET_OFFICE_URL = "https://backend.rentify.rbit.pe/api/offices/";
-  static const String GET_OFFICES_URL = "https://backend.rentify.rbit.pe/api/offices";
+  static const String GET_RANDOM_USER_LIST_URL =
+      "https://randomuser.me/api"; //https://randomuser.me/api
+  static const String GET_RESERVATION_LIST_URL =
+      "https://rentifyupc.azurewebsites.net/api/reservations"; //http://localhost:5000/api/reservations - https://localhost:5001/api/reservations
+  static const String GET_OFFICE_URL =
+      "https://rentifyupc.azurewebsites.net/api/offices/";
+  static const String GET_OFFICES_URL =
+      "https://rentifyupc.azurewebsites.net/api/offices";
 }
