+++ lib/main.dart
@@ -1,11 +1,10 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:rentify/views/create_reservation.dart';
-import 'package:rentify/views/login.dart';
+import 'package:rentify/views/auth/login.dart';
 import './navigation/router.dart';
 
 void main() async {
-  runApp(MyApp(initRoute: '/login'));
+  runApp(MyApp(initRoute: '/app'));
 }
 
 class MyApp extends StatelessWidget {
@@ -20,12 +19,8 @@ class MyApp extends StatelessWidget {
       initialRoute: this.initRoute,
       routes: <String, WidgetBuilder>{
         //'/authentication': (BuildContext context) => AuthenticationApp(),
-        '/app': (BuildContext context) =>
-            AppRouter(),
-        '/create_reservation': (BuildContext context) =>
-            CreateReservationPage(),
-        '/login': (BuildContext context) =>
-            Login(),
+        '/app': (BuildContext context) => AppRouter(),
+        '/login': (BuildContext context) => Login(),
       },
     );
   }
