+++ lib/views/office/detail_office.dart
