+++ lib/views/office/list_my_offices.dart
@@ -1,0 +1,7 @@
+/*
+*
+* Es una pagina que te permite ver las oficinas que has creado.
+* Tambien incluye el boton "Create workplace". Si le das click a
+*  una de tus oficinas en la lista, te lleva a la otra Ruta:
+* detalle de oficina (provider)
+* */
