+++ lib/views/office/list_office_reservations.dart
@@ -1,0 +1,7 @@
+/*
+* Es una pagina donde puedes ver las reservas que se han hecho  sobre tu oficina,
+* a modo de card list. Tiene 3 secciones para ver reservas activas (aceptadas),
+* inactivas (denegadas o que ya terminaron), y solicitud de reservas (pending)
+*
+*
+* */
