--- lib/views/settings.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/cupertino.dart';
-
-class SettingsPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Text('Settings Page'),
-    );
-  }
-}
